requirements.txt
pywebview>=4.0.2
numpy>=1.22.0
requests>=2.27.1